parte 2

+++ nahuel/src/main/java/com/unla/nahuel/util/ListaUsuariosPdf.java
@@ -131,7 +131,7 @@ public class ListaUsuariosPdf extends AbstractPdfView {
 				celda.setPadding(5);
 				tablaLista.addCell(celda);
 				
-				Integer documento = usuario.getDocumento();
+				Long documento = usuario.getDocumento();
 				celda = new PdfPCell(new Phrase(documento.toString(), fuenteDatos));
 				celda.setPadding(5);
 				tablaLista.addCell(celda);
